Move `NetworkConfiguration` to `node` module
