+++ exonum/src/events/network.rs
@@ -37,9 +37,8 @@ use crate::{
         error::into_failure,
         noise::{Handshake, HandshakeParams, NoiseHandshake},
     },
-    helpers::Milliseconds,
     messages::{Connect, Message, Service, SignedMessage, Verified},
-    node::SharedConnectList,
+    node::{NetworkConfiguration, SharedConnectList},
 };
 use exonum_crypto::x25519::into_x25519_public_key;
 
@@ -74,30 +73,6 @@ pub enum NetworkRequest {
     SendMessage(PublicKey, SignedMessage),
     DisconnectWithPeer(PublicKey),
     Shutdown,
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone, Copy, PartialEq)]
-pub struct NetworkConfiguration {
-    // TODO: Think more about config parameters. (ECR-162)
-    pub max_incoming_connections: usize,
-    pub max_outgoing_connections: usize,
-    pub tcp_nodelay: bool,
-    pub tcp_keep_alive: Option<u64>,
-    pub tcp_connect_retry_timeout: Milliseconds,
-    pub tcp_connect_max_retries: u64,
-}
-
-impl Default for NetworkConfiguration {
-    fn default() -> Self {
-        Self {
-            max_incoming_connections: 128,
-            max_outgoing_connections: 128,
-            tcp_keep_alive: None,
-            tcp_nodelay: true,
-            tcp_connect_retry_timeout: 15_000,
-            tcp_connect_max_retries: 10,
-        }
-    }
 }
 
 #[derive(Debug)]
